ssz: Move to own pkg
